Commit hashes rendered on toolWindow get consumed from Service
